requirements.txt
numpy==1.15.4
pandas==0.23.4
mecab-python3==0.996.1
mojimoji==0.0.8
scipy==1.1.0
Flask==1.0.2
scikit-learn==0.20.1
lightgbm==2.2.1
optuna==0.7.0
